tlsn::connection

Enum CertBinding

#[non_exhaustive]
pub enum CertBinding {
    V1_2(CertBindingV1_2),
}

TLS certificate binding.

This is the data that the server signs using its public key in the certificate it presents during the TLS handshake. This provides a binding between the serverโs identity and the ephemeral keys used to authenticate the TLS session.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

V1_2(CertBindingV1_2)

TLS 1.2 certificate binding.