tlsn::transcript::encoding

Trait Encoder

pub trait Encoder {
    // Required methods
    fn encode_range(
        &self,
        direction: Direction,
        range: Range<usize>,
        dest: &mut Vec<u8>,
    );
    fn encode_data(
        &self,
        direction: Direction,
        range: Range<usize>,
        data: &[u8],
        dest: &mut Vec<u8>,
    );
}

A transcript encoder.

This is an internal implementation detail that should not be exposed to the public API.

Required Methods

fn encode_range( &self, direction: Direction, range: Range<usize>, dest: &mut Vec<u8>, )

Writes the zero encoding for the given range of the transcript into the destination buffer.

fn encode_data( &self, direction: Direction, range: Range<usize>, data: &[u8], dest: &mut Vec<u8>, )

Writes the encoding for the given data into the destination buffer.